+++ skbio/alignment/tests/test_tabular_msa.py
@@ -210,15 +210,13 @@ class TestTabularMSA(unittest.TestCase, ReallyEqualMixin):
         self.assertIsInstance(msa.index, pd.MultiIndex)
         assert_index_equal(msa.index, pd.Index([('foo', 42), ('bar', 43)]))
 
-    def test_copy_constructor_handles_missing_metadata_efficiently(self):
+    def test_copy_constructor_without_metadata(self):
         msa = TabularMSA([DNA('ACGT'), DNA('----')])
 
         copy = TabularMSA(msa)
 
-        self.assertIsNone(msa._metadata)
-        self.assertIsNone(msa._positional_metadata)
-        self.assertIsNone(copy._metadata)
-        self.assertIsNone(copy._positional_metadata)
+        self.assertEqual(msa, copy)
+        self.assertIsNot(msa, copy)
 
     def test_copy_constructor_with_metadata(self):
         msa = TabularMSA([DNA('ACGT'),
@@ -2234,6 +2232,17 @@ class TestExtend(unittest.TestCase):
             TabularMSA([DNA('AA'),
                         DNA('GG')], index=['foo', 'bar']))
 
+    def test_extending_on_empty_with_positional_metadata(self):
+        # bug in 0.4.2
+        msa = TabularMSA([], positional_metadata={'foo': []})
+
+        msa.extend([DNA('AA'), DNA('GG')])
+
+        self.assertEqual(
+            msa,
+            TabularMSA([DNA('AA'),
+                        DNA('GG')]))
+
 
 class TestJoin(unittest.TestCase):
     def test_invalid_how(self):
@@ -2265,7 +2274,7 @@ class TestJoin(unittest.TestCase):
             TabularMSA([DNA('AC'), DNA('--')]).join(
                 TabularMSA([DNA('GT'), DNA('..')], index=[0, 0]))
 
-    def test_handles_missing_metadata_efficiently(self):
+    def test_no_metadata(self):
         msa1 = TabularMSA([DNA('AC'),
                            DNA('G.')])
         msa2 = TabularMSA([DNA('-C'),
@@ -2277,12 +2286,6 @@ class TestJoin(unittest.TestCase):
             joined,
             TabularMSA([DNA('AC-C'),
                         DNA('G..G')]))
-        self.assertIsNone(msa1._metadata)
-        self.assertIsNone(msa1._positional_metadata)
-        self.assertIsNone(msa2._metadata)
-        self.assertIsNone(msa2._positional_metadata)
-        self.assertIsNone(joined._metadata)
-        self.assertIsNone(joined._positional_metadata)
 
     def test_ignores_metadata(self):
         msa1 = TabularMSA([DNA('AC', metadata={'id': 'a'}),
@@ -2670,16 +2673,6 @@ class TestIterPositions(unittest.TestCase):
                       positional_metadata={'foo': [42, np.nan, -1],
                                            'bar': [np.nan, np.nan, 'baz']})])
 
-    def test_handles_missing_positional_metadata_efficiently(self):
-        msa = TabularMSA([DNA('AC'),
-                          DNA('A-')])
-
-        self.assertIsNone(msa._positional_metadata)
-
-        list(msa.iter_positions())
-
-        self.assertIsNone(msa._positional_metadata)
-
 
 class TestConsensus(unittest.TestCase):
     def test_no_sequences(self):
@@ -2771,17 +2764,6 @@ class TestConsensus(unittest.TestCase):
             cons,
             DNA('A-T', positional_metadata={'foo': [42, 43, 42],
                                             'bar': ['a', 'b', 'c']}))
-
-    def test_handles_missing_positional_metadata_efficiently(self):
-        msa = TabularMSA([DNA('AC'),
-                          DNA('AC')])
-
-        self.assertIsNone(msa._positional_metadata)
-
-        cons = msa.consensus()
-
-        self.assertIsNone(msa._positional_metadata)
-        self.assertIsNone(cons._positional_metadata)
 
     def test_mixed_gap_characters_as_majority(self):
         seqs = [
@@ -3353,14 +3335,6 @@ class TestGetPosition(unittest.TestCase):
 
         self.assertEqual(position,
                          Sequence('C-', metadata={'foo': 43, 'bar': 'def'}))
-
-    def test_handles_positional_metadata_efficiently(self):
-        msa = TabularMSA([DNA('AA'),
-                          DNA('--')])
-
-        msa._get_position_(1)
-
-        self.assertIsNone(msa._positional_metadata)
 
 
 class TestIsSequenceAxis(unittest.TestCase):